Use dynamic pkgver

+++ PKGBUILD
@@ -1,50 +1,59 @@
 # Maintainer: Kaizhao Zhang <zhangkaizhao@gmail.com>
 
 _pkgname=gobeansdb
-_gitcommit=7127da29f8e9a4312ee600271e98b29ab6771753
 
-pkgname=gobeansdb-git
-pkgver=0.0.0+2+7127da2
+pkgname="${_pkgname}-git"
+pkgver=r623.d940fae
 pkgrel=1
 pkgdesc="High performance distributed object storage server from Douban Inc."
 url="https://github.com/douban/gobeansdb"
-makedepends=('git' 'go' 'vgo-git')
+makedepends=('go>=1.11')
 options=('!strip' '!emptydirs')
 arch=('x86_64')
 license=('BSD-3-Clause')
-source=("git+${url}.git#commit=${_gitcommit}")
+conflicts=('gobeansdb')
+source=("git+${url}.git")
 sha256sums=('SKIP')
+
+pkgver() {
+  cd "${_pkgname}"
+  printf "r%s.%s" "$(git rev-list --count HEAD)" "$(git rev-parse --short HEAD)"
+}
 
 prepare() {
   cd "${srcdir}"
 
-  rm -rf "${srcdir}/goext"
+  if [[ -d "${srcdir}/goext" ]]; then
+    # Why read-only? :/
+    chmod -R u=rwX,go=rX "${srcdir}/goext"
+
+    rm -rf "${srcdir}/goext"
+  fi
 }
 
 build() {
-  cd "${srcdir}"
+  cd "${srcdir}/${_pkgname}"
 
   export GOPATH="${srcdir}/goext"
+  #export GO111MODULE=on
 
-  install -d -m755 "${GOPATH}/src/github.com/douban"
-  ln -s "${srcdir}/${_pkgname}" "${GOPATH}/src/github.com/douban/gobeansdb"
-
-  cd "${GOPATH}/src/github.com/douban/gobeansdb"
-  vgo install github.com/douban/gobeansdb
+  go install ./
 }
 
 package() {
   install -d -m755 "${pkgdir}"/usr/bin
   install -d -m755 "${pkgdir}"/usr/share/"${_pkgname}"/{bin,conf,tests}
 
-  # binary file
   export GOPATH="${srcdir}/goext"
   cd "${GOPATH}"
+
+  # binary file
   install -D -m755 bin/gobeansdb "${pkgdir}/usr/share/${_pkgname}/bin/gobeansdb"
   ln -s "/usr/share/${_pkgname}/bin/gobeansdb" "${pkgdir}/usr/bin/gobeansdb"
 
-  # conf and tests files
   cd "${srcdir}/${_pkgname}"
+
+  # conf and tests files
   install -D -m644 conf/* "${pkgdir}/usr/share/${_pkgname}/conf/"
   install -D -m644 tests/* "${pkgdir}/usr/share/${_pkgname}/tests/"
